Refactor filtering of x_values in Function class
This commit refactors the filtering of x_values in the Function class of math_geometry.py. Instead of using a list comprehension, a try-except block is used to handle ZeroDivisionError. This ensures that x_values where f(x) results in a ZeroDivisionError are skipped. The filtered values are then appended to the filtered_values list. This change improves the robustness of the code.